refactor: reorganise YAML processor into dedicated module with phase-based naming
Major reorganisation for better code structure and clarity:

File Structure Changes:
- Create data_model/yaml_processor/ subdirectory for all YAML processing
- Rename files to reflect three-phase pipeline design:
  - uptodate_yaml.py → phase_a_parameter_update.py
  - science_check.py → phase_b_science_check.py
  - phase_c_reports.py → phase_c_pydantic_report.py
  - suews_yaml_processor.py → orchestrator.py
  - precheck.py → validation_helpers.py

Test Organisation:
- Move YAML processor tests to test/data_model/yaml_processor/
- Keep core data model tests in test/data_model/

Documentation:
- Add comprehensive README.md in yaml_processor directory
- Developer documentation now co-located with code
- Clear explanation of three-phase pipeline architecture

Benefits:
- Clearer separation between data definitions and processing logic
- Phase-based naming makes pipeline structure self-documenting
- Tests organised to match source structure
- Developer docs maintained alongside code
