+++ src/supy/data_model/yaml_processor/__init__.py
@@ -1,0 +1,23 @@
+"""
+SUEWS YAML Processor Module
+
+This module contains a three-phase processing pipeline for SUEWS YAML configuration files.
+
+Pipeline Phases:
+- Phase A: Parameter detection and updating (missing/renamed parameters)
+- Phase B: Scientific validation (physics constraints, model dependencies)
+- Phase C: Pydantic validation and reporting
+
+Components:
+- phase_a_parameter_update: Phase A implementation
+- phase_b_science_check: Phase B implementation  
+- phase_c_pydantic_report: Phase C implementation
+- orchestrator: Pipeline orchestrator for running phases
+- validation_helpers: Shared validation utilities
+"""
+
+from .phase_a_parameter_update import *
+from .phase_b_science_check import *
+from .phase_c_pydantic_report import *
+from .orchestrator import *
+from .validation_helpers import *
