+++ src/supy/meson.build
@@ -73,11 +73,12 @@ py.install_sources(
             'data_model/validation_utils.py',
             'data_model/yaml_annotator.py',
             'data_model/yaml_annotator_json.py',
-            'data_model/precheck.py',
-            'data_model/uptodate_yaml.py',
-            'data_model/science_check.py',
-            'data_model/suews_yaml_processor.py',
-            'data_model/phase_c_reports.py',
+            'data_model/yaml_processor/__init__.py',
+            'data_model/yaml_processor/phase_a_parameter_update.py',
+            'data_model/yaml_processor/phase_b_science_check.py',
+            'data_model/yaml_processor/phase_c_pydantic_report.py',
+            'data_model/yaml_processor/orchestrator.py',
+            'data_model/yaml_processor/validation_helpers.py',
         ],
         [
             'ext_data/CRU_TS4.06_1991_2020.parquet',
