+++ test/data_model/yaml_processor/__init__.py
@@ -1,0 +1,1 @@
+"""Test suite for SUEWS YAML Processor components."""
